Create invertTree
I made a mistake here. I created a helper function within the invertTree function. And was using node1 and node2 in it. I only needed one argument. So the proper way to resolve it is to create a invertTree function and call the helper function from there (only one argument in helper function)